requirements.txt
Flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
Werkzeug>=3.0.0
gunicorn==21.2.0
